requirements.txt
django==4.1.2
psycopg2-binary==2.9.4
django-debug-toolbar==3.7.0
 djangorestframework==3.14.0